Home

RequirementVersions.jl

RequirementVersions.copy_with_permissions

RequirementVersions.fix_permissions

RequirementVersions.minimum_requirement_versions

RequirementVersions.restore

RequirementVersions.minimum_requirement_versions — Method.

minimum_requirement_versions(package_name; package_directory = Pkg.dir(), should_resolve = true, skips = String[])

Automatically finds the minimum versions of required packages that will still allow your tests to pass. List any packages you want to skip in skips. Set should_resolve to false to prevent Pkg.pin and Pkg.test from running Pkg.resolve().

Makes the assumption that if a certain profile of versions work, all profiles with versions greater or equal will also work.

source

RequirementVersions.copy_with_permissions — Method.

copy_with_permissions(source, destination; remove_destination = false)

Copy source to destination, preserving permissions

source

RequirementVersions.fix_permissions — Method.

fix_permissions(source, destination)

Ensure destination has same permissions as source

source

RequirementVersions.restore — Function.

restore(archive, package_directory = Pkg.dir())

Restore all the files in the archive to your package directory.

source